% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{annotate}
\alias{annotate}
\title{Annotate ontology with gene annotations.}
\usage{
annotate(o, annotation)
}
\arguments{
\item{o}{An \code{ontology} object.}

\item{annotations}{A named list of character vectors.}
}
\value{
An \code{ontology} object with annotations.
}
\description{
Annotate ontology with gene annotations.
}
