% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{collapse_redundant_terms}
\alias{collapse_redundant_terms}
\title{Collapse terms whose annotations are redundant with respect to their children.}
\usage{
collapse_redundant_terms(o)
}
\arguments{
\item{o}{An annotated \code{ontology} object.}
}
\value{
A filtered \code{ontology} object.
}
\description{
Collapse terms whose annotations are redundant with respect to their children.
}
