% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{collapse_terms}
\alias{collapse_terms}
\title{Collapse a set of terms, rerouting their children directly to their parents.}
\usage{
collapse_terms(o, terms)
}
\arguments{
\item{o}{An \code{ontology} object.}

\item{terms}{A character vectors of terms to collapse.}
}
\description{
Collapse a set of terms, rerouting their children directly to their parents.
}
