% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{compute_redundancy}
\alias{compute_redundancy}
\title{Compute redundancy score for each term.}
\usage{
compute_redundancy(o)
}
\arguments{
\item{o}{An annotated \code{ontology} object.}
}
\value{
A numeric vector of Jaccard index values (between 0 and 1) for each term.
    Terms for which the score is not defined (e.g. terms with no children or no annotations) will have an \code{NA} value.
}
\description{
The score is defined as the maximum Jaccard index between the annotations of a term and the annotations of its children.
}
