% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.R
\name{compute_term_pcs}
\alias{compute_term_pcs}
\title{Compute principal components of SNPs annotated to ontology terms.}
\usage{
compute_term_pcs(
  path,
  o,
  data,
  terms = NULL,
  stand = "binom2",
  max_term_size = 75,
  min_term_size = 3,
  explain_var = 0.95,
  max_pcs = 50,
  num_parts = 1,
  exclude_snps = NULL,
  rsvd_threshold = Inf
)
}
\arguments{
\item{path}{Path to output file.}

\item{o}{An annotated \code{ontology} object.}

\item{data}{A \code{BEDData} object}

\item{terms}{A character vector of terms to compute PCs for. Will use all terms if not provided.}

\item{stand}{Which standardization method to use. One of "none", "binom" (old Eigenstrat-style), "binom2" (new Eigenstrat-style), "sd" (zero-mean unit-variance) or "center" (zero mean).}

\item{max_term_size}{Skip terms annotated with fewer than this number of genes.}

\item{explain_var}{Restrict number of PCs to explain at least this fraction of variance.}

\item{max_pcs}{Return at most this number of PCs.}

\item{num_parts}{Number of parts to write output to.}

\item{exclude_snps}{Named list of SNP IDs to exclude from terms.}

\item{rsvd_threshold}{Use randomized SVD when number of variables exceeds this threshold. Set to \code{Inf} to turn off.}
}
\description{
Compute principal components of SNPs annotated to ontology terms.
}
