% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.R
\name{extract_pcs}
\alias{extract_pcs}
\title{Extract the PCs of a specific set of terms.}
\usage{
extract_pcs(path, terms)
}
\arguments{
\item{path}{Path to PC file.}

\item{terms}{Character vector of terms (or genes) to extract.}
}
\description{
Extract the PCs of a specific set of terms.
}
