% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{filter_obsolete}
\alias{filter_obsolete}
\title{Removes obsolete terms from the ontology.}
\usage{
filter_obsolete(o, keep_connected = TRUE)
}
\arguments{
\item{o}{An \code{ontology} object.}

\item{keep_connected}{Keep obsolete terms that are ancestors of non-obsolete terms.}
}
\value{
An ontology with all obsolete terms removed.
}
\description{
Removes obsolete terms from the ontology.
}
