% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{filter_unannotated}
\alias{filter_unannotated}
\title{Removes unannotated terms from the ontology.}
\usage{
filter_unannotated(o, keep_connected = TRUE)
}
\arguments{
\item{o}{An annotated \code{ontology} object.}

\item{keep_connected}{Keep unannotated terms that are ancestors of annotated terms.}
}
\value{
An ontology with all unannotated terms removed.
}
\description{
Removes unannotated terms from the ontology.
}
