% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{get_genemap}
\alias{get_genemap}
\title{Map genes to SNPs.}
\usage{
get_genemap(snps, ref = "hg19", maxgap = 10000)
}
\arguments{
\item{snps}{A}

\item{ref}{Reference genome. One of "hg18", "hg19", "hg38".}

\item{maxgap}{Maximum allowed distance between SNP and gene.}
}
\description{
Map genes to SNPs.
}
