% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{make_genemap}
\alias{make_genemap}
\title{Create mapping from genes to SNPs.}
\usage{
make_genemap(snps, ref = "hg19", maxgap = 10000)
}
\arguments{
\item{snps}{A}

\item{ref}{Reference genome. One of "hg18", "hg19", "hg38".}

\item{maxgap}{Maximum allowed distance between SNP and gene.}
}
\value{
A data frame mapping genes to SNPs.
}
\description{
Create mapping from genes to SNPs.
}
