% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot part of an ontology as a tree.}
\usage{
plot(o, terms, include_ancestors = TRUE, pvalues = NULL)
}
\arguments{
\item{o}{An \code{ontology} object.}

\item{terms}{Terms to include in plot.}

\item{include_ancestors}{Include the ancestors of terms in the plot.}

\item{pvalues}{A named numeric vector of p-values for each term. Term nodes will be colored according to -log10(p-value) if provided.}
}
\description{
Plot part of an ontology as a tree.
}
