% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hierarchy}
\alias{plot_hierarchy}
\title{Plot part of an ontology as a tree.}
\usage{
plot_hierarchy(path, o, terms, test = NULL, include_ancestors = TRUE,
  highlight = character(0), highlight.color = "black",
  text_wrap = 20, font = "Arial")
}
\arguments{
\item{path}{Path to save Graphviz DOT file.}

\item{o}{An \code{ontology} object.}

\item{terms}{Character vector of terms to include in the plot.}

\item{test}{Test result object computed with \code{\link{test_terms}}. Used to color terms according to p-value.}

\item{include_ancestors}{Include all ancestors of terms in `terms`.}

\item{highlight}{Character vector of terms to highlight.}

\item{highlight.color}{Border color for highlighted terms.}

\item{text_wrap}{Wrap term descriptions at this line width.}

\item{font}{Name of font to use in nodes.}
}
\value{
Path to produced graph.
}
\description{
Plot part of an ontology as a tree.
}
