% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{propagate_annotations}
\alias{propagate_annotations}
\title{Propagates annotations up through the hierarchy.
After propagation the set of annotations for a term is a superset of the
union of annotations for its children.}
\usage{
propagate_annotations(o)
}
\arguments{
\item{o}{An \code{ontology} object.}
}
\value{
A propagated \code{ontology} object.
}
\description{
Propagates annotations up through the hierarchy.
After propagation the set of annotations for a term is a superset of the
union of annotations for its children.
}
