% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_bed}
\alias{read_bed}
\title{Open PLINK .bed file and create gene mapping.}
\usage{
read_bed(path, ref = "hg19", maxgap = 10000, select = "all")
}
\arguments{
\item{path}{Path to .bed file with or without extension.}

\item{ref}{Reference genome. One of "hg18", "hg19", "hg38", "mm8", "mm9", "mm10", "rn5", "rn6".}

\item{maxgap}{Maximum allowed distance between SNP and gene.}

\item{select}{How should SNPs near multiple genes by mapped? Use "all" to use all hits. Use "nearest" to use only nearest gene.}
}
\value{
A \code{BEDData} object.
}
\description{
Open PLINK .bed file and create gene mapping.
}
