% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{read_gaf}
\alias{read_gaf}
\title{Read annotations from GAF file.}
\usage{
read_gaf(path, aspects, exclude_evidence)
}
\arguments{
\item{path}{Path to GAF file.}

\item{aspects}{Character vector of aspects to include. Leave blank to include all aspects.}

\item{exclude_evidence}{Character vector of evidence types to exclude.}
}
\description{
Read annotations from GAF file.
}
