% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{recompute_ancestors}
\alias{recompute_ancestors}
\title{Recomputes \code{ancestors} field for each term.}
\usage{
recompute_ancestors(o)
}
\arguments{
\item{o}{An \code{ontology} object.}
}
\description{
This method can be used to repair the ontology after changing the \code{children} or \code{parents} fields.
}
