% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{test_genes}
\alias{test_genes}
\title{Perform association test for genes.}
\usage{
test_genes(formula, covars, pc, npcs = Inf)
}
\arguments{
\item{formula}{An object of class \code{formula}.}

\item{covars}{A data frame containing covariates.}

\item{pc}{Named list of PCs from genes. Computed with \code{\link{compute_gene_pcs}}.}

\item{npcs}{Number of PCs to use. Set to \code{Inf} to use all available.}
}
\value{
A list with elements:
  \item{\code{test}}{Test objects for each gene.}
  \item{\code{pvalue}}{p-values for each gene.}
}
\description{
Perform association test for genes.
}
