% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{test_terms_children}
\alias{test_terms_children}
\title{Check terms for significance against children.}
\usage{
test_terms_children(formula, covars, pc, o, family = binomial("logit"),
  terms = NULL, max_pcs = Inf, num_threads = 1)
}
\arguments{
\item{formula}{An object of class \code{formula}.}

\item{covars}{A data frame containing covariates.}

\item{pc}{Named list of PCs from terms. Computed with \code{\link{compute_term_pcs}}.}

\item{o}{An \code{ontology} object.}

\item{family}{Error distribution and link function to be used in the model. See \code{\link{glm.fit}} for details.}

\item{terms}{Character vector of terms to test. Defaults to all terms.}

\item{max_pcs}{Maximum number of PCs to use per term.}

\item{num_threads}{Number of threads.}
}
\value{
A list with elements:
  \describe{
    \item{\code{tests}}{Test objects for each term.}
    \item{\code{pvalue}}{p-values for each term.}
  }
}
\description{
Check terms for significance against children.
}
