% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{unite_roots}
\alias{unite_roots}
\title{Unites a set of roots in the ontology under a common root.
Primarily used to combine multiple aspects of an ontology,
e.g. 'biological process', 'cellular component' and 'molecular function' in Gene Ontology.}
\usage{
unite_roots(o, roots, root.name = "root")
}
\arguments{
\item{o}{An \code{ontology} object.}

\item{roots}{A character vector of roots to combine. Will unite all roots in ontology if not provided.}

\item{root.name}{Name of the new root term.}
}
\value{
The united \code{ontology} object.
}
\description{
Unites a set of roots in the ontology under a common root.
Primarily used to combine multiple aspects of an ontology,
e.g. 'biological process', 'cellular component' and 'molecular function' in Gene Ontology.
}
