% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.R
\name{compute_term_pcs}
\alias{compute_term_pcs}
\title{Compute principal components of SNPs annotated to ontology terms.}
\usage{
compute_term_pcs(o, geno, genemap, npcs = 4, terms = NULL,
  max_term_size = Inf, stand = "binom2", progress = interactive(),
  ...)
}
\arguments{
\item{o}{An annotated \code{ontology} object.}

\item{geno}{A \code{SnpMatrix} genotype matrix.}

\item{genemap}{A data frame mapping genes to SNP rs numbers.}

\item{npcs}{Number of principal components to compute per term.}

\item{terms}{A character vector of terms to compute PCs for. Will use all terms if not provided.}

\item{max_term_size}{Skip terms annotated with more than this number of genes.}

\item{stand}{Which standardization method to use. One of "none", "binom" (old Eigenstrat-style), "binom2" (new Eigenstrat-style), "sd" (zero-mean unit-variance) or "center" (zero mean).}

\item{progress}{Show progress bar?}

\item{...}{Further arguments to \code{\link{flashpcaR::flashpca}}.}
}
\value{
A list of matrices where each column corresponds to a PC.
}
\description{
Compute principal components of SNPs annotated to ontology terms.
}
