% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.R
\name{compute_term_pcs}
\alias{compute_term_pcs}
\title{Compute principal components of SNPs annotated to ontology terms.}
\usage{
compute_term_pcs(o, geno, genemap, npcs = 4, max_term_size = Inf)
}
\arguments{
\item{o}{An annotated \code{ontology} object.}

\item{geno}{A \code{SnpMatrix} genotype matrix.}

\item{genemap}{A data frame mapping genes to SNP rs numbers.}

\item{npcs}{Number of principal components to compute per term.}

\item{max_term_size}{Skip terms annotated with more than this number of genes.}
}
\description{
Compute principal components of SNPs annotated to ontology terms.
}
