% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.R
\name{compute_term_pcs}
\alias{compute_term_pcs}
\title{Compute principal components of SNPs annotated to ontology terms.}
\usage{
compute_term_pcs(o, geno, genemap, stand = "binom2", terms = NULL,
  max_term_size = 75, min_term_size = 3, explain_var = 0.95,
  max_pcs = 50, exclude_snps = NULL, rsvd_threshold = 0)
}
\arguments{
\item{o}{An annotated \code{ontology} object.}

\item{geno}{A \code{SnpMatrix} genotype matrix.}

\item{genemap}{A data frame mapping genes to SNP rs numbers.}

\item{stand}{Which standardization method to use. One of "none", "binom" (old Eigenstrat-style), "binom2" (new Eigenstrat-style), "sd" (zero-mean unit-variance) or "center" (zero mean).}

\item{terms}{A character vector of terms to compute PCs for. Will use all terms if not provided.}

\item{max_term_size}{Skip terms annotated with more than this number of genes.}

\item{explain_var}{Restrict number of PCs to explain at least this fraction of variance.}

\item{max_pcs}{Return at most this number of PCs.}

\item{exclude_snps}{Named list of SNP IDs to exclude from terms.}

\item{rsvd_threshold}{Use randomized SVD when number of variables exceeds this threshold.}

\item{max_term_size}{Skip terms annotated with fewer than this number of genes.}
}
\value{
A list of singular value decompositions for each term. Each entry is a list with elements
  \describe{
    \item{\code{u}}{The left-singular vectors.}
    \item{\code{d}}{The non-zero singular values.}
    \item{\code{v}}{The right-singular vectors.}
  }
}
\description{
Compute principal components of SNPs annotated to ontology terms.
}
