% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{filter_obsolete}
\alias{filter_obsolete}
\title{Removes obsolete terms from the ontology.}
\usage{
filter_obsolete(o, keep_connected = TRUE)
}
\arguments{
\item{o}{An \code{ontology} object.}

\item{keep_connected}{Don't remove ancestors of terms in \code{terms}.}
}
\value{
The filtered ontology.
}
\description{
Removes obsolete terms from the ontology.
}
