% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{permute}
\alias{permute}
\title{Permutes the hierarchy of an ontology while preserving the in- and out-degree of each terms.}
\usage{
permute(o)
}
\arguments{
\item{An}{\code{ontology} object.}
}
\description{
Permutes the hierarchy of an ontology while preserving the in- and out-degree of each terms.
}
