% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hierarchy}
\alias{plot_hierarchy}
\title{Plot part of an ontology as a tree.}
\usage{
plot_hierarchy(o, path, terms, pvalues = NULL,
  include_ancestors = TRUE, highlight = character(0), text_wrap = 20,
  font = "Arial")
}
\arguments{
\item{o}{An \code{ontology} object.}

\item{path}{Path to save Graphviz DOT file.}

\item{terms}{Character vector of terms to include in the plot.}

\item{pvalues}{Named numeric vector of p-values to use for coloring nodes.}

\item{include_ancestors}{Include all ancestors of terms in `terms`.}

\item{highlight}{Character vector of terms to highlight.}

\item{text_wrap}{Wrap term descriptions at this line width.}

\item{font}{Name of font to use in nodes.}
}
\value{
Path to produced graph.
}
\description{
Plot part of an ontology as a tree.
}
