% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{test_genes}
\alias{test_genes}
\title{Perform association test for genes.}
\usage{
test_genes(formula, covars, pc, family = binomial("logit"),
  num_threads = 1)
}
\arguments{
\item{formula}{An object of class \code{formula}.}

\item{covars}{A data frame containing covariates.}

\item{pc}{Named list of PCs from genes. Computed with \code{\link{compute_gene_pcs}}.}

\item{family}{Error distribution and link function to be used in the model. See \code{\link{glm.fit}} for details.}

\item{progress}{Show progress bar?}
}
\value{
A list with elements:
  \item{\code{test}}{Test objects for each gene.}
  \item{\code{pvalue}}{p-values for each gene.}
}
\description{
Perform association test for genes.
}
