% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{test_genes}
\alias{test_genes}
\title{Perform association test for genes.}
\usage{
test_genes(
  path,
  formula,
  covars,
  family = binomial("logit"),
  test = NULL,
  max_pcs = Inf
)
}
\arguments{
\item{path}{Path to PC data.}

\item{formula}{An object of class \code{formula}.}

\item{covars}{A data frame containing covariates.}

\item{family}{Error distribution and link function to be used in the model. See \code{\link{glm.fit}} for details.}

\item{test}{Test statistic for model comparison. One of "Chisq", "LRT", "Rao", "F" or "Cp". See \code{\link{stat.anova}}.}

\item{max_pcs}{Maximum number of PCs to use per gene.}
}
\value{
A list with elements:
  \describe{
    \item{\code{test}}{Data frame of test statistics.}
    \item{\code{coef}}{Coefficients from each term model.}
  }
}
\description{
Perform association test for genes.
}
