% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{test_terms}
\alias{test_terms}
\title{Perform association test for terms.}
\usage{
test_terms(formula, covars, pc, family = binomial("logit"),
  num_threads = 1)
}
\arguments{
\item{formula}{An object of class \code{formula}.}

\item{covars}{A data frame containing covariates.}

\item{pc}{Named list of PCs from terms. Computed with \code{\link{compute_term_pcs}}.}

\item{family}{Error distribution and link function to be used in the model. See \code{\link{glm.fit}} for details.}

\item{o}{An \code{ontology} object.}

\item{progress}{Show progress bar?}
}
\value{
A list with elements:
  \describe{
    \item{\code{reference}}{Reference model.}
    \item{\code{coef}}{Coefficients from each term model.}
    \item{\code{tests}}{Test objects for each term.}
    \item{\code{pvalue}}{p-values for each term.}
  }
}
\description{
Perform association test for terms.
}
