% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{test_terms}
\alias{test_terms}
\title{Perform association test for all terms}
\usage{
test_terms(formula, covars, pc, npcs)
}
\arguments{
\item{formula}{An object of class \code{formula}.}

\item{covars}{A data frame containing covariates.}

\item{pc}{Named list of PCs from terms. Computed with \code{\link{compute_term_pcs}}.}

\item{npcs}{Number of PCs to use.}
}
\description{
Perform association test for all terms
}
