% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{test_terms}
\alias{test_terms}
\title{Perform association test for terms.}
\usage{
test_terms(formula, covars, pc, family = binomial("logit"),
  test = NULL, terms = NULL, max_pcs = Inf, num_threads = 1)
}
\arguments{
\item{formula}{An object of class \code{formula}.}

\item{covars}{A data frame containing covariates.}

\item{pc}{Named list of PCs from terms. Computed with \code{\link{compute_term_pcs}}.}

\item{family}{Error distribution and link function to be used in the model. See \code{\link{glm.fit}} for details.}

\item{terms}{Character vector of terms to test. Defaults to all terms.}

\item{max_pcs}{Maximum number of PCs to use per term.}

\item{num_threads}{Number of threads.}

\item{Test}{statistic for model comparison. One of "Chisq", "LRT", "Rao", "F" or "Cp". See \code{\link{stat.anova}}.}
}
\value{
A list with elements:
  \describe{
    \item{\code{test}}{Test objects for each term.}
    \item{\code{coef}}{Coefficients from each term model.}
  }
}
\description{
Perform association test for terms.
}
