% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{test_terms}
\alias{test_terms}
\title{Perform association test for all terms}
\usage{
test_terms(formula, covars, pc, o = NULL, npcs = Inf,
  against = "whole.genome")
}
\arguments{
\item{formula}{An object of class \code{formula}.}

\item{covars}{A data frame containing covariates.}

\item{pc}{Named list of PCs from terms. Computed with \code{\link{compute_term_pcs}}.}

\item{o}{An \code{ontology} object.}

\item{npcs}{Number of PCs to use. Set of \code{Inf} to use all available.}

\item{against}{Model to test against for significance. One of "whole.genome", "parent".}
}
\value{
A list with elements:
  \item{\code{test}}{Test objects for each term.}
  \item{\code{pvalue}}{p-values for each term.}
}
\description{
Perform association test for all terms
}
